+++ src/const.js
@@ -1,0 +1,7 @@
+export const STORAGES_LIST_ROUTE = '/settings/remote-storage';
+
+export const SECTIONS_STORAGE = {
+  INFORMATION: 'information',
+  ACCESSION: 'accession',
+  RETRIEVAL: 'retrieval',
+};