+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Stories
 * [UIRS-4](https://issues.folio.org/browse/UIRS-4) Project Setup: ui-remote-storage
 * [UIRS-3](https://issues.folio.org/browse/UIRS-3) Permission: Remote storage configuration
+* [UIRS-3](https://issues.folio.org/browse/UIRS-2) Remote storage configuration